Refactor frontend into modular structure and improve streaming utils
