+++ frontend/index.html
@@ -274,6 +274,6 @@
       </section>
     </template>
 
-    <script type="module" src="app.js"></script>
+    <script type="module" src="app/index.js"></script>
   </body>
 </html>
